rename: 관리자에서만 사용하는 컴포넌트, 상수, 타입 관리자로 이동
