+++ src/pages/Admin/Assignment/components/AssignmentTable.tsx
@@ -1,8 +1,9 @@
 import { useEffect, useState } from 'react';
 
-import { tableColumnMap, tableColumns } from '@/constants/adminTableColumns';
 import useModal from '@/hooks/useModal';
-import { AssignmentTableData } from '@/types/admin';
+
+import { tableColumnMap, tableColumns } from '../../constants';
+import { AssignmentTableData } from '../../types';
 
 import AssignmentDetailModal from './AssignmentDetailModal';
